DTS-39282: moving to 10.1.0
